rbac final

+++ crates/services/web-server/src/web/rest/drug_recurrence_rest.rs
@@ -3,6 +3,10 @@
 use axum::extract::{Path, State};
 use axum::http::StatusCode;
 use axum::Json;
+use lib_core::model::acs::{
+	DRUG_RECURRENCE_CREATE, DRUG_RECURRENCE_DELETE, DRUG_RECURRENCE_LIST,
+	DRUG_RECURRENCE_READ, DRUG_RECURRENCE_UPDATE,
+};
 use lib_core::model::drug_recurrence::{
 	DrugRecurrenceInformation, DrugRecurrenceInformationBmc,
 	DrugRecurrenceInformationFilter, DrugRecurrenceInformationForCreate,
@@ -11,7 +15,7 @@ use lib_core::model::drug_recurrence::{
 use lib_core::model::ModelManager;
 use lib_rest_core::rest_params::{ParamsForCreate, ParamsForUpdate};
 use lib_rest_core::rest_result::DataRestResult;
-use lib_rest_core::Result;
+use lib_rest_core::{require_permission, Result};
 use lib_web::middleware::mw_auth::CtxW;
 use modql::filter::{ListOptions, OpValValue, OpValsValue};
 use serde_json::json;
@@ -26,6 +30,7 @@ pub async fn create_drug_recurrence(
 	Json(params): Json<ParamsForCreate<DrugRecurrenceInformationForCreate>>,
 ) -> Result<(StatusCode, Json<DataRestResult<DrugRecurrenceInformation>>)> {
 	let ctx = ctx_w.0;
+	require_permission(&ctx, DRUG_RECURRENCE_CREATE)?;
 	tracing::debug!(
 		"{:<12} - rest create_drug_recurrence drug_id={}",
 		"HANDLER",
@@ -53,6 +58,7 @@ pub async fn list_drug_recurrences(
 	Json<DataRestResult<Vec<DrugRecurrenceInformation>>>,
 )> {
 	let ctx = ctx_w.0;
+	require_permission(&ctx, DRUG_RECURRENCE_LIST)?;
 	tracing::debug!(
 		"{:<12} - rest list_drug_recurrences drug_id={}",
 		"HANDLER",
@@ -85,6 +91,7 @@ pub async fn get_drug_recurrence(
 	Path((_case_id, _drug_id, id)): Path<(Uuid, Uuid, Uuid)>,
 ) -> Result<(StatusCode, Json<DataRestResult<DrugRecurrenceInformation>>)> {
 	let ctx = ctx_w.0;
+	require_permission(&ctx, DRUG_RECURRENCE_READ)?;
 	tracing::debug!("{:<12} - rest get_drug_recurrence id={}", "HANDLER", id);
 
 	let entity = DrugRecurrenceInformationBmc::get(&ctx, &mm, id).await?;
@@ -101,6 +108,7 @@ pub async fn update_drug_recurrence(
 	Json(params): Json<ParamsForUpdate<DrugRecurrenceInformationForUpdate>>,
 ) -> Result<(StatusCode, Json<DataRestResult<DrugRecurrenceInformation>>)> {
 	let ctx = ctx_w.0;
+	require_permission(&ctx, DRUG_RECURRENCE_UPDATE)?;
 	tracing::debug!("{:<12} - rest update_drug_recurrence id={}", "HANDLER", id);
 
 	let ParamsForUpdate { data } = params;
@@ -118,6 +126,7 @@ pub async fn delete_drug_recurrence(
 	Path((_case_id, _drug_id, id)): Path<(Uuid, Uuid, Uuid)>,
 ) -> Result<StatusCode> {
 	let ctx = ctx_w.0;
+	require_permission(&ctx, DRUG_RECURRENCE_DELETE)?;
 	tracing::debug!("{:<12} - rest delete_drug_recurrence id={}", "HANDLER", id);
 
 	DrugRecurrenceInformationBmc::delete(&ctx, &mm, id).await?;
